re-structure generators folder
